perf(page-load): lazy-init DashboardView deep-link import hooks
DashboardView called five Firestore-listener hooks at the top of every
teacher mount — useQuiz, useQuizAssignments, useVideoActivity,
useVideoActivityAssignments, usePlcs — but consumed ONLY their import/peek
callbacks (never their assignments/quizzes/plcs lists; the Quiz and
VideoActivity widgets run their own independent listeners). So every normal
teacher load paid for five live onSnapshot subscriptions plus the associated
callback/effect setup to service a code path that only matters when a share
link is actually being imported.

Extracts that machinery (the five hooks, the open*WidgetToTab helpers, the
three pending-share effects, runAssignmentImport / peekAndDispatchImport /
runVideoActivityImport, and the import mode-picker modal) into a new
DeepLinkShareImporter, mounted lazily and only once a pending share id
appears. The mount is latched (imports clear their pending id synchronously,
so a bare presence gate would unmount mid-import) and behaviour is identical:
the effects fire on mount with the same pending id, clear it synchronously to
avoid the triple-import race, and surface the same toasts / picker.

Net effect on the common teacher load: zero of these Firestore listeners are
opened (a real cost reduction not visible to the mocked perf harness), and the
import-machinery callbacks/effects no longer run at mount.

No new Firestore/Storage/AI usage — strictly fewer listeners on the common
path; the import paths issue exactly the same reads/writes as before.
